IOS-5033 Moved formatting to viewModel
