+++ Tangem/Modules/Send/Summary/SendWalletSummaryViewModel.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 class SendWalletSummaryViewModel: Identifiable {
     let walletName: String
@@ -15,5 +16,43 @@ class SendWalletSummaryViewModel: Identifiable {
     init(walletName: String, totalBalance: String) {
         self.walletName = walletName
         self.totalBalance = totalBalance
+    }
+
+    func walletNameTitle(fontSize: CGFloat) -> NSAttributedString {
+        return TangemRichTextFormatter().format(Localization.sendFromWallet(walletName), fontSize: fontSize)
+    }
+}
+
+// MARK: - Rich text formatter
+
+private struct TangemRichTextFormatter {
+    // Formatting rich text as NSAttributedString
+    // Supported formats: **bold**
+    func format(_ string: String, fontSize: CGFloat) -> NSAttributedString {
+        var originalString = string
+
+        let regex = try! NSRegularExpression(pattern: "\\*{2}.+?\\*{2}")
+
+        let wholeRange = NSRange(location: 0, length: (originalString as NSString).length)
+        let matches = regex.matches(in: originalString, range: wholeRange)
+
+        let attributedString = NSMutableAttributedString(string: originalString)
+
+        if let match = matches.first {
+            let formatterTagLength = 2
+
+            let boldTextFormatted = String(originalString[Range(match.range, in: originalString)!])
+            let boldText = boldTextFormatted.dropFirst(formatterTagLength).dropLast(formatterTagLength)
+
+            originalString = originalString.replacingOccurrences(of: boldTextFormatted, with: boldText)
+            attributedString.setAttributedString(NSAttributedString(string: originalString))
+
+            // UIKit's .semibold corresponds SwiftUI bold font
+            let boldFont = UIFont.systemFont(ofSize: fontSize, weight: .semibold)
+            let boldTextRange = NSRange(location: match.range.location, length: match.range.length - 2 * formatterTagLength)
+            attributedString.addAttribute(.font, value: boldFont, range: boldTextRange)
+        }
+
+        return attributedString
     }
 }
